requirements.txt
# Dependências do Opportunity Bot
python-socketio==5.8.0
aiohttp==3.8.5
discord.py==2.3.2
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
